fix(security): move unconditional security beans to auto-config; relocate @EnableWebSecurity
methodSecurityExpressionHandler, httpSessionEventPublisher, and authenticationEventPublisher were
@Bean methods on the component-scanned WebSecurityConfig with no @ConditionalOnMissingBean, so a
consumer-defined bean of the same type caused an override conflict. They move to
UserSecurityBeansAutoConfiguration (each now guarded by @ConditionalOnMissingBean), consistent with
the codebase principle that @ConditionalOnMissingBean is only reliable on auto-configuration classes.

@EnableWebSecurity moves from WebSecurityConfig to WebSecurityFilterChainAutoConfiguration, where the
filter chain is actually defined. The auto-config Javadoc now warns prominently that its type-based
@ConditionalOnMissingBean back-off is all-or-nothing: ANY consumer SecurityFilterChain bean (even a
narrow one) suppresses the library chain entirely.

Addresses review findings HIGH-3, HIGH-4, and MEDIUM-1.
